moving ship component out of sidebar again...........
